Move InMemoryKeyValueStorage out of production source into test target
Test-only stub doesn't need to ship in the PointOfSale module. Moves the class verbatim into Modules/Tests/PointOfSaleTests/Roles/Mocks/. The protocol it conforms to (POSStaffKeyValueStorage) stays internal in production; tests reach it via @testable import PointOfSale.

Mirrors the existing test-mock convention (WooCommerce/WooCommerceTests/Mocks/MockTableView.swift et al.).
